bug fixed & added button Load More & category modification

+++ controllers/categoryCtrl.js
@@ -1,5 +1,6 @@
 
 const Category = require('../models/categoryModel')
+const Products = require('../models/productModel')
 
 const categoryCtrl = {
 getCategories: async (req, res) => {
@@ -30,6 +31,9 @@ createCategory: async (req, res) => {
 },
 deleteCategory: async (req, res) =>{
     try {
+        const products = await Products.findOne({category: req.params.id})
+        if(products) return res.status(455).json({msg: "Please delete all products with a relationship."})
+
         await Category.findByIdAndDelete(req.params.id)
         res.json({msg: 'Category deleted'})
     } catch (err) {
